Switch to i32 for handles. Finish "aclose".

+++ ailets-rs/actor_runtime/src/actor_runtime.rs
@@ -3,7 +3,7 @@ extern "C" {
     pub fn n_of_streams(name_ptr: *const i8) -> i32;
     pub fn open_read(name_ptr: *const i8, index: usize) -> i32;
     pub fn open_write(name_ptr: *const i8) -> i32;
-    pub fn aread(fd: usize, buffer_ptr: *mut u8, count: usize) -> i32;
-    pub fn awrite(fd: usize, buffer_ptr: *const u8, count: usize) -> i32;
-    pub fn aclose(fd: usize) -> i32;
+    pub fn aread(fd: i32, buffer_ptr: *mut u8, count: usize) -> i32;
+    pub fn awrite(fd: i32, buffer_ptr: *const u8, count: usize) -> i32;
+    pub fn aclose(fd: i32) -> i32;
 }
